remove unused simple_gcluster_p()
