+++ src/lib/internal.h
@@ -1,11 +1,13 @@
 #ifndef NOTCURSES_INTERNAL
 #define NOTCURSES_INTERNAL
 
+#include <term.h>
 #include <stdio.h>
 #include <stdint.h>
 #include <stdarg.h>
 #include <string.h>
 #include <stdbool.h>
+#include <pthread.h>
 #include "egcpool.h"
 
 #ifdef __cplusplus
@@ -63,6 +65,52 @@ typedef struct ncvisual {
   int stream_index;     // match against this following av_read_frame()
   ncplane* ncp;
 } ncvisual;
+
+typedef struct notcurses {
+  pthread_mutex_t lock;
+  int ttyfd;      // file descriptor for controlling tty, from opts->ttyfp
+  FILE* ttyfp;    // FILE* for controlling tty, from opts->ttyfp
+  FILE* ttyinfp;  // FILE* for processing input
+  int colors;     // number of colors usable for this screen
+  ncstats stats;  // some statistics across the lifetime of the notcurses ctx
+  // We verify that some terminfo capabilities exist. These needn't be checked
+  // before further use; just use tiparm() directly.
+  char* cup;      // move cursor
+  char* civis;    // hide cursor
+  char* clear;    // erase screen and home cursor
+  // These might be NULL, and we can more or less work without them. Check!
+  char* cnorm;    // restore cursor to default state
+  char* smcup;    // enter alternate mode
+  char* rmcup;    // restore primary mode
+  char* setaf;    // set foreground color (ANSI)
+  char* setab;    // set background color (ANSI)
+  char* op;       // set foreground and background color to default
+  char* standout; // CELL_STYLE_STANDOUT
+  char* uline;    // CELL_STYLE_UNDERLINK
+  char* reverse;  // CELL_STYLE_REVERSE
+  char* blink;    // CELL_STYLE_BLINK
+  char* dim;      // CELL_STYLE_DIM
+  char* bold;     // CELL_STYLE_BOLD
+  char* italics;  // CELL_STYLE_ITALIC
+  char* italoff;  // CELL_STYLE_ITALIC (disable)
+  char* smkx;     // enter keypad transmit mode (keypad_xmit)
+  char* rmkx;     // leave keypad transmit mode (keypad_local)
+
+  // special keys
+  char* left;  // kcub1
+  char* right; // kcuf1
+  char* up;    // kcuu1
+  char* down;  // kcud1
+  char* npage; // knp
+  char* ppage; // kpp
+
+  struct termios tpreserved; // terminal state upon entry
+  bool RGBflag;   // terminfo-reported "RGB" flag for 24bpc directcolor
+  bool CCCflag;   // terminfo-reported "CCC" flag for palette set capability
+  ncplane* top;   // the contents of our topmost plane (initially entire screen)
+  ncplane* stdscr;// aliases some plane from the z-buffer, covers screen
+  FILE* renderfp; // debugging FILE* to which renderings are written
+} notcurses;
 
 #ifdef __cplusplus
 }
